converter, storage: extract DeviceNamer
Move DeviceNamer and device naming logic from converter.go to
converter/storage/devicenaming.go
