+++ pkg/virt-launcher/notify-client/BUILD.bazel
@@ -16,7 +16,7 @@ go_library(
         "//pkg/virt-launcher/virtwrap/agent-poller:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//pkg/virt-launcher/virtwrap/cli:go_default_library",
-        "//pkg/virt-launcher/virtwrap/converter:go_default_library",
+        "//pkg/virt-launcher/virtwrap/converter/storage:go_default_library",
         "//pkg/virt-launcher/virtwrap/errors:go_default_library",
         "//pkg/virt-launcher/virtwrap/storage:go_default_library",
         "//pkg/virt-launcher/virtwrap/util:go_default_library",
