+++ pkg/virt-launcher/virtwrap/converter/storage/BUILD.bazel
@@ -3,6 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
+        "devicenaming.go",
         "directio.go",
         "generated_mock_directio.go",
         "paths.go",
@@ -15,6 +16,7 @@ go_library(
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//pkg/virtiofs:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
+        "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//vendor/go.uber.org/mock/gomock:go_default_library",
     ],
 )
@@ -22,6 +24,7 @@ go_library(
 go_test(
     name = "go_default_test",
     srcs = [
+        "devicenaming_test.go",
         "directio_test.go",
         "storage_suite_test.go",
         "virtiofs_test.go",
@@ -31,6 +34,7 @@ go_test(
         ":go_default_library",
         "//pkg/libvmi:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
+        "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
